added mesa headers and tounix.sh script
i added the 'tounix.sh' script to the tree root which normalizes line endings,
filemodes and strips trailing withspaces from source files.
furthermore i added the current mesa headers (7.1) to 'libsrc/mesa-7.1' which
will be only used when building the scenery tools on linux using the Makefile
under the tree root. the reason is pretty simple. when installing the linux
nvidia drivers the nvidia installer overwrites the current mesa headers which
is absolutetly the correct behaviour to ensure that we only use extensions
which are implemented and supported by the current driver.
the problem is that these headers doesn't contain some extension declarations
like glClientActiveTextureARB() while they are perfectly supported by the
driver (the extensions which caused the compilation problems are opengl 1.2
extensions). we can safely use the mesa headers in this scenario but when
messing with extensions introduced with opengl 2.0 or newer we have to be
careful, since some of them might not be implemented in the current nvidia
driver yet (in which case we can simply switch back using standard driver
headers by modifying the makefile).