added projectfiles for visual studio 2008
including a fix for buildsettings of b64.c (i changed this one to have really
c code). building it with the old project files might cause an unresolved
symbol error (_decode). i have no visual studio 2005 and was too lazy to fix
this manually in the project files. any chance you can take care of it ben? :-)
the new solution also contains a freetype project, so building it in an extra
step with the shipped project files isn't necessary anymore. this commit
also includes a small fix in two headers of the GUI framework, since msvc
doesn't support the #warning preprocessor directive :-(