project file fixes
removed deprecated configurations. also renamed Persistance.cp to
Persistance.cpp and updated xcode project file and linux makefile.
Ben, if anything else is affected by this rename please let me know. it
is possible to define a custom build step in visual studio 2008 for this
but visual studio saves this information in a user file, not in the project
file itself, which is not what i call useful. again the old visual studio 2005
files need to be changed to work again after this fix :-)