add mingw compiler support
building on windows with gcc 4.3.3 works. WED segfaults on startup,
need to investigate this later, the rest seem to work fine. once
we have a public git repository availiable, i'll prepare a convenient
installerpackage for windows, which will bootstrap the build environment
(compressed around 10MB for the i386 architecture) and the git repository
from the internet, so that one can build the scenery tools easily on the
windows platform or start hacking on it without any additional dependencies.

Signed-off-by: Janos Laube <janos.dev@gmail.com>