WED, win32: fix crash on startup
Inconsistent typedefs in libjpeg caused inproper structure sizes
which in turn led to an unhandled exception. This however only occured
when building WED with gcc on Windows, which has Boolean defined as
unsigned char in a system header, while libjpeg defines it as int per
default. Working around this by typedef'ing Boolean to unsigned char
in jmorecfg.h.
Also fixed a bug in BitmapUtils: jpeg_create_decompress() can throw
exceptions, by calling setjmp() before that, we have the possibility
to catch them.
Furthermore added new tri stripper files to the makerules which makes
the tree buildable again on Windows with gcc and Linux.

Signed-off-by: Janos Laube <janos.dev@gmail.com>