allow use of prebuilt 3rd-party libraries
Introduce new make variable 'gitlibs'. When setting this to '1' make will clone
a git repository with prebuild libraries and put it into ./libs/local instead
of building the libraries from scratch. due to binary compatibility
inconsistencies on posix compatible systems we support this for windows only.
The repository is ~16MB and will be updated when ever i update the tarball
system counterpart. This should speed up the library build step on windows
massively.

Signed-off-by: Janos Laube <janos.dev@gmail.com>