wed, linux: mouse related fixes by Mathias
I have changed the mouse eventshandler,because they causes a bug .
(internal button-numbers can not assigned twice ).
Also a issue is that some XButtons hold the down-state till click it again.
Since all Xbuttons do the same as Midbutton ,
It makes a confused because you do not know which key is responsible and
 how can be turned off.
I suggest not to send XButton until it really needed from 'WED'

also implemented 'Changing CursorShapes' .

Written-by: Mathias Roedel <m_roedel@web.de>
Signed-off-by: Janos Laube <janos.dev@gmail.com>