linux: add LD_RUN_PATH configuration
This gives us support for shipping in-tree shared libraries for official
linux binary distributions of the Scenery Tools.

With the current setup binaries should behave like on Windows, i.e. the
dynamic loader will search first in the directory of the executable, then
in ./lib (ok, this is not like on Windows :-)) and then the default system
directories for dependencies. This even allows us to ship libc as shared
object, thus working around limitations of statical linked glibc
applications. The only dependency which can't be resolved that way is the
dynamic loader itself (typically /lib/ld-linux.so), which means that we
need at least glibc 2.3 on the host system when compiling the Scenery
Tools with a recent toolchain.

Signed-off-by: Janos Laube <janos.dev@gmail.com>