linux, wed: initial drag&drop support
According to Mathias this one might need to be revised:

"i've started with implementing D&D functionality for QT.
There are one main-problem  -> detecting 'start-drag' .
While Ben calls OS-functions and wait of there result ,we must do this
with our own input-handlers .

In my possible solution , i call the event-loop in a own loop till a
drag is detected or Mouse-button is released.
While Drag-detection , all 'ClickUp'-Calls to the GUI_Window  blocked.
And at the end  , i send the missing 'UPClick'"

Written-by: Mathias Roedel <m_roedel@web.de>
Signed-off-by: Janos Laube <janos.dev@gmail.com>