makefile fixes for mac os
we can't link statically against libgcc anymore, since the ppc-version of
the 10.4 sdk in xcode 3.2 was compiled with strange stack protector
settings.
also xcode 3.2 makes it impossible to compile debug versions of the tools
for the ppc architecture. the linker dies with 'bl range out of bounds'
error, which seems to mean that the branch instruction only can branch to
an offset of 16MB at max. try to compile with -mlong-branch? but xcode
3.2 says that this is not needed anymore :-S.
best to drop ppc support, hehe.

Signed-off-by: Janos Laube <janos.dev@gmail.com>