CVS Resync.  There is a bug in the CVS to GIT migration that can cause branches to have incorrect file checkouts when the branch is tag isn't consistent in "time" on CVS (since CVS has non-atomic checkins).
This apparently is what was going on on the MT branch - the GIT MT 2.0 maintenance branch has MUCH newer code than the CVS one, which is really bad if we want to do a MT 2.x maintenance release.

So..this commit "blasts" the "src" directory with a direct checkout from CVS, resetting the files, but not really re-creating MT's history.  Since we expect to do maybe one MT 2.x maintenance release, this will have to work for now.