Added comments through out various classes. I also put in a functionality that stops an previously unknown bug.
The Bug: earth.wed.xml exists but is without content), however it still trys to create a Document with it.
The fix: In document if it gets a document amounting to "" it throws an error.
Problems: It stops any other document from opening.
I also came up with a way to hide the console if you don't need it.