"added the single line "mEditInfo.text_val = it->second.first.c_str();", 638
Inside of filter bar is the fully working functionality of having a popup menu containing all the packs inside of the package manager. This is done with the lines 146 and 148, perfected and cleaned GetEnumDictionary, and the additional lines inside AcceptEdit "if(cell_x == 1 && cell_y == 0) and etc".

there is still the bug that you must click twice for the text to appear.