This is a big commit. By Class:
LLA: Added mCurIntVal, added a parameter to SetFilter, Refactored Rebuild Cache to Work with the new enum system (this refactoring is mainly the large switch statement, move the block of code dictating where mCatLoc/LibInd are to the DoFilter Method. In the header are some updates to the commenting.

FilterBar: Updated GetCellContent to handle the new enum section (see the new switch statement lines 146 to 155), forced the enum dictionary to have "Local", "Library", and "All" choices (matching the enum values located in the Library Manager).

GUI_Messages: Removed a useless enum from when this filtering feature was going to be a splitter.