Awesome commit! Starting from the top,
MemFileHandling, fixed up error messages so they are better for the user
TimerFired, making it so that if any errors happen in the import process the operation is canceled and we go to the error window
FillVersionsFromJSON, the all important one's for the column are given defaults so the sorting of them doesn't break. the default of them is to ensure that a null answer always has is less than a real date. If there is no accepted or approved date the pack is not pushed into the vector of versions
ImportSpecificVersion, commenting and error message change, nothing special
DecorateGUIWindow, ensures that whenever something is downloading bother tables stay hidden