bug fix: fixed runaway timer in hierarchy
The hierarchy drag timer was running indefinitely and pushing bogus
drag events to the tools because the timer stop code for Mac was
borked.  This would cause a zero length drag in the bezier tool,
causing it to “miss” its drag.

The vertex tool’s code is pretty awful  - it’s evolved more and more
stageful complexity and this kind of hard to find bug is the result,
but for now, fixing the timer fixes the bug AND the timer API was
100% broken so it has to be fixed.

fixes WED-559