fix illegal memory access typing in password field of gateway upload.
This is a really hard bug to see without ASAN - it happens when the
memory at 'end()' of the char vector used to store the password
***** version of typing is accessed by the unicode next-codepoint
finder.  Normally vector either has a little extra memory, or it's
heap allocated and you can read the malloc guard page without exploding.

With asan, the read off the end of the vector (which is technically
illegal) triggers an immediate hard crash, revealing the bug.  To
see it, turn on asan and type non-unicode into the gateway upload
password field.

The fix provides a range-based version of the unicode char finder
that honors the 'next' pointer.