Improve result of approx_t_for_xy() for curve segments close to vertical.
Previously, the parameter value t that was returned was always one of the
roots that was computed in x-direction. This could be inaccurate at parts of
the curve that were almost vertical.

Instead, choose whichever of the two roots (in x- or y-direction) gets us
closer to the desired point. This makes the result noticeably more accurate
at vertical parts of the curve.