refactor (mobile autogen terrain): Prep for treating terrain selection as a pipeline
We work with the terrain types *just* in terms of their grid (x, y) coordinates for an arbitrary number of steps;
only at the end do we actually assign the final terrain enums to the Pmwx.

This is a helluva lot more straightforward than iterating the Pmwx faces multiple times.