Change Slippy Map MAX_ZOOM to 21.
I want to be able to use higher resolution imagery when using a custom slippy map. I found that the provider I was using could provide down to z23. The WED map window only allows the user to zoom to z21. The MAX_ZOOM should be updated to match for custom providers offering high resolution data. Also could make it user customization in the preferences pane.