feat (mobile AG terrain; XPM-1272): Initial attempt at fixing floating buildings
This is not great in European AG---it removes more than it has to, and it fails to remove some things it *should*. The issue here is that each Euro OBJ is actually a *block* of buildings, so a 3 m delta from the far left to the far of a 100 m block produces noticable floating. This is going to require an art fix (either breaking up the blocks into their constituent buildings, or adding a "basement" to all the buildings such that some amount of floating gets papered over).