Renamed the command line option to take --gateway_api_url="https://gatewayapi.x-plane.com:3001/apiv1/"
and added an option to overrride the built in cert via
--gateway_cert="path_to_cert"

Its neccessary, as cURL will find and bomb upon any mismatch of the
built-ini cert URL and whatever its being overridden with.
As the current production and test servers both respond to
--gateway_api_url="http://gatewaytest.x-plane.com/apiv1/"
as well, that may not be needed unless actually submiting something.