WED roads : temporär fix for import export network shape nodes
In the current interpretion network shape points have no height info in the DSF.
But we need to provide them for WED.
Also by export the converting of 2-way nodes to shape points or junction nodes is unpredicted.
Until a better solution , shape nodes get -999 for in and out level to indentify
and also rexport them as such.