Add PixelSize() methods to WED_MapZoomerNew.
These provide a more general way of determining the pixel size of objects
on screen than the fixed "pixel per meter" value provided by GetPPM()
(which doesn't work in the case of a perspective projection).

The methods either forward to the corresponding PixelSize() methods of
the WED_Camera, if one was provided, or use the GetPPM() value if no
camera was provided.