Add an options struct to WED_PreviewLayer.
Some entities need to be drawn slightly differently in the map window
and the preview window. The options struct is used to control this.

Options are stored in WED_PreviewItem in a new member variable "option_bits".
Note that on a 64-bit build, the size of WED_PreviewItem has not
changed. It was 16 bytes prior to this change (due to alignment
requirements) with 4 unused bytes of padding. The new member variable
option_bits uses those 4 previously unused bytes.