Move clearing of depth buffer from WED_PreviewLayer to WED_Map.
This ensures that WED_PreviewLayer does not overwrite the z buffer
values of the ground plane drawn by WED_MapPreviewPane. The z values of
the ground plane are needed to prevent the parts of building basements
that are under ground from being drawn.

All layers in the map that come before WED_PreviewLayer turn the z buffer
off, so this has no effect on the rendering of the map.