Make line placement visibility decision respect minLineThicknessPixels.
This is necessary to make line placements show correctly in the preview
window and matches what we do for airport chains.

As a consequence, we have to move the code that draws a "shadow" for
line placements if we're zoomed too far out to WED_StructureLayer. This
is desirable for the preview window too, as we don't want to show the
"shadow" there.